Update RSpec config files
