+++ .rspec
@@ -1,1 +1,2 @@
---colour
+--color
+--require spec_helper
