+++ spec/support/utils_helper.rb
@@ -1,0 +1,10 @@
+module UtilsHelper
+  def login!
+    visit '/users/sign_in'
+    within('#new_user') do
+      fill_in 'Email', with: @user.email
+      fill_in 'Password', with: 'rubytw8888'
+    end
+    click_button 'Sign in'
+  end
+end
